requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.8.1
ipfshttpclient==0.8.0a2
isort==5.6.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.7
mccabe==0.6.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.18.1
py-solc==3.2.0
py-solc-x==1.1.1
pycryptodome==3.11.0
pylint==2.6.0
pyrsistent==0.18.0
python-dotenv==0.19.1
requests==2.26.0
rlp==2.0.1
semantic-version==2.8.5
six==1.12.0
toml==0.10.2
toolz==0.11.1
typed-ast==1.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
wrapt==1.12.1
yarl==1.7.0
zipp==3.6.0
